+++ plugin-dev/Source/Sentry/Private/Desktop/SentrySubsystemDesktop.h
@@ -49,6 +49,8 @@ public:
 
 	void TryCaptureScreenshot() const;
 
+	FString GetGpuDumpBackupPath() const;
+
 	TSharedPtr<SentryScopeDesktop> GetCurrentScope();
 
 private: